Avoid FrozenError when handler returns a frozen string
When calling an external command and reading the output, `io.read`
returns `""` when no output is available, but `io.read(max_size)`
returns `nil` instead of `""`.

This return value is then converted to a string with `to_s`, and since
ruby 2.7, `nil.to_s` returns a frozen empty string. Then the
`Plaintext::Resolver` calling `gsub!` and `strip!` on this string lead
to a `FrozenError: can't modify frozen String`.

This is fixed by duplicating the string if it is frozen.